+++ R/RcppExports.R
@@ -25,6 +25,10 @@
     .Call(`_roll_roll_median`, x, width, weights, min_obs, complete_obs, na_restore, online)
 }
 
+.roll_max <- function(x, width, weights, min_obs, complete_obs, na_restore, online) {
+    .Call(`_roll_roll_max`, x, width, weights, min_obs, complete_obs, na_restore, online)
+}
+
 .roll_var <- function(x, width, weights, center, min_obs, complete_obs, na_restore, online) {
     .Call(`_roll_roll_var`, x, width, weights, center, min_obs, complete_obs, na_restore, online)
 }