+++ R/roll.R
@@ -228,6 +228,46 @@ roll_median <- function(x, width, weights = rep(1, width),
   ))
 }
 
+##' Rolling Maximums
+##'
+##' A function for computing rolling maximums of time-series data.
+##'
+##' @param x matrix or xts object. Rows are observations and columns are variables.
+##' @param width integer. Window size.
+##' @param weights vector. Weights for each observation within a window.
+##' @param min_obs integer. Minimum number of observations required to have a value within a window,
+##' otherwise result is \code{NA}.
+##' @param complete_obs	logical. If \code{TRUE} then rows containing any missing values are removed,
+##' if \code{FALSE} then each value is used.
+##' @param na_restore logical. Should missing values be restored?
+##' @param online logical. Process observations using an online algorithm.
+##' @return An object of the same class and dimension as \code{x} with the rolling maximums
+##' @examples
+##' n_vars <- 3
+##' n_obs <- 15
+##' x <- matrix(rnorm(n_obs * n_vars), nrow = n_obs, ncol = n_vars)
+##' 
+##' # rolling maximums
+##' result <- roll_max(x, 5)
+##' 
+##' # rolling maximums with exponential decay
+##' weights <- 0.9 ^ (5:1)
+##' result <- roll_max(x, 5, weights)
+##' @export
+roll_max <- function(x, width, weights = rep(1, width),
+                     min_obs = width, complete_obs = FALSE, na_restore = FALSE,
+                     online = FALSE) {
+  return(.Call(`_roll_roll_max`,
+               x,
+               as.integer(width),
+               as.numeric(weights),
+               as.integer(min_obs),
+               as.logical(complete_obs),
+               as.logical(na_restore),
+               as.logical(online)
+  ))
+}
+
 ##' Rolling Variances
 ##'
 ##' A function for computing rolling variances of time-series data.