+++ src/init.c
@@ -4,14 +4,15 @@
 #include <R_ext/Rdynload.h>
 
 /* FIXME: 
- Check these declarations against the C/Fortran source code.
- */
+Check these declarations against the C/Fortran source code.
+*/
 
 /* .Call calls */
 extern SEXP _roll_roll_all(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 extern SEXP _roll_roll_any(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 extern SEXP _roll_roll_cov(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 extern SEXP _roll_roll_lm(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
+extern SEXP _roll_roll_max(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 extern SEXP _roll_roll_mean(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 extern SEXP _roll_roll_median(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 extern SEXP _roll_roll_prod(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
@@ -25,6 +26,7 @@ static const R_CallMethodDef CallEntries[] = {
   {"_roll_roll_any",    (DL_FUNC) &_roll_roll_any,     6},
   {"_roll_roll_cov",    (DL_FUNC) &_roll_roll_cov,    10},
   {"_roll_roll_lm",     (DL_FUNC) &_roll_roll_lm,      9},
+  {"_roll_roll_max",    (DL_FUNC) &_roll_roll_max,     7},
   {"_roll_roll_mean",   (DL_FUNC) &_roll_roll_mean,    7},
   {"_roll_roll_median", (DL_FUNC) &_roll_roll_median,  7},
   {"_roll_roll_prod",   (DL_FUNC) &_roll_roll_prod,    7},