+++ NEWS.md
@@ -2,7 +2,8 @@
 
 ### Changes in roll version 1.1.2 (YYYY-MM-DD)
 
-* New `roll_median`, `roll_any`, and `roll_all` functions for computing rolling medians, any, and all, respectively, of time-series data ([#4](https://github.com/jjf234/roll/issues/4), [#14](https://github.com/jjf234/roll/issues/14))
+* New `roll_median`, `roll_max`, `roll_any`, and `roll_all` functions for computing rolling medians, maximums, any, and all, respectively, of time-series data ([#4](https://github.com/jjf234/roll/issues/4), [#13](https://github.com/jjf234/roll/issues/13), [#14](https://github.com/jjf234/roll/issues/14))
+    * Note: `roll_median` and `roll_max` are not calculated using online algorithms
 
 ### Changes in roll version 1.1.1 (2018-08-11)